Moved SpellInfo to seperated file
