+++ src/world/Spell/CMakeLists.txt
@@ -11,6 +11,8 @@ set(SRC_SPELL_FILES
    ${PATH_PREFIX}/SpellDefines.hpp
    ${PATH_PREFIX}/SpellEffects.cpp
    ${PATH_PREFIX}/SpellFailure.h
+   ${PATH_PREFIX}/SpellInfo.cpp
+   ${PATH_PREFIX}/SpellInfo.hpp
    ${PATH_PREFIX}/SpellMgr.cpp
    ${PATH_PREFIX}/SpellMgr.h
    ${PATH_PREFIX}/SpellNameHashes.h
