+++ src/world/Spell/SpellInfo.cpp
@@ -1,0 +1,6 @@
+/*
+Copyright (c) 2014-2016 AscEmu Team <http://www.ascemu.org/>
+This file is released under the MIT license. See README-MIT for more information.
+*/
+
+#include "StdAfx.h"
